api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]>=0.31.1

# SQLModel (統合ライブラリ)
sqlmodel==0.0.14

# Database drivers
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Migrations
alembic==1.13.1

# Settings
pydantic-settings>=2.1.0

# MCP (Model Context Protocol)
fastmcp>=2.0.0,<3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx>=0.27.0

# Test data factories
faker==22.0.0
factory-boy==3.3.0
